+++ bin/KGbuilder/KGbuilder_triplesMaker.js
@@ -159,10 +159,17 @@ var KGbuilder_triplesMaker = {
             return null;
         }
         var p = mappingValue.indexOf("]");
+
         if (p > 0) {
             //specific baseURI
             var baseUri = mappingValue.substring(1, p);
             var value = line[mappingValue.substring(p + 1)];
+            var columnURI = mappingValue.substring(p + 1);
+            var prefixURI = tableMappings.prefixURI[columnURI];
+            if (prefixURI) {
+                baseUri += prefixURI;
+            }
+
             // specific base URI can also have transforms and lookups
             if (value) {
                 if (tableMappings.transform && tableMappings.transform[mappingValue]) {
@@ -244,15 +251,19 @@ var KGbuilder_triplesMaker = {
                 return callback((lineError = e + " " + mapping.s));
             }
         } else if (typeof mapping.s === "string" && mapping.s.indexOf("http") == 0) {
-            subjectStr = "<" + mapping.s + ">";
-        } else if (typeof mapping.s === "string" && mapping.s.match(/.+:.+/)) {
+            var prefixURI = tableMappings.prefixURI[mapping.s] || "";
+            subjectStr = "<" + prefixURI + mapping.s + ">";
+        }
+        // sparql prefix
+        else if (typeof mapping.s === "string" && mapping.s.match(/.+:.+/)) {
             subjectStr = mapping.s;
         } else {
             if (!line[mapping.s] || (mapping.o.indexOf(":") > -1 && line[mapping.o]) == "null") {
                 return callback(null, null);
             }
+            var prefixURI = tableMappings.prefixURI[mapping.s] || "";
 
-            subjectStr = line[mapping.s];
+            subjectStr = prefixURI + line[mapping.s];
         }
         if (mapping.lookup_s && !isTransformLookUp) {
             if (!lookUpsMap[mapping.lookup_s]) {
@@ -262,7 +273,8 @@ var KGbuilder_triplesMaker = {
             if (!lookupValue) {
                 missingLookups_s += 1;
             } else {
-                subjectStr = lookupValue;
+                var prefixURI = tableMappings.prefixURI[mapping.s] || "";
+                subjectStr = prefixURI + lookupValue;
                 okLookups_s += 1;
             }
         }
@@ -280,7 +292,9 @@ var KGbuilder_triplesMaker = {
         } else if (KGbuilder_triplesMaker.isUri(subjectStr)) {
             subjectStr = "<" + subjectStr + ">";
         } else {
-            subjectStr = "<" + tableMappings.graphUri + util.formatStringForTriple(subjectStr, true) + ">";
+            var prefixURI = tableMappings.prefixURI[mapping.s] || "";
+
+            subjectStr = "<" + prefixURI + tableMappings.graphUri + util.formatStringForTriple(subjectStr, true) + ">";
         }
 
         return callback(null, subjectStr);
@@ -315,7 +329,8 @@ var KGbuilder_triplesMaker = {
                 return callback(e);
             }
         } else if (typeof mapping.o === "string" && mapping.o.indexOf("http") == 0) {
-            objectStr = "<" + mapping.o + ">";
+            var prefixURI = tableMappings.prefixURI[mapping.o] || "";
+            objectStr = "<" + prefixURI + mapping.o + ">";
             return callback(null, objectStr);
         } else if (typeof mapping.o === "string" && mapping.o.match(/.+:.+/)) {
             objectStr = mapping.o;
@@ -433,7 +448,8 @@ var KGbuilder_triplesMaker = {
                 objectStr = '"' + str + '"^^' + mapping.dataType;
             } else {
                 if (isTransform == false) {
-                    objectStr = line[mapping.o];
+                    var prefixURI = tableMappings.prefixURI[mapping.o] || "";
+                    objectStr = prefixURI + line[mapping.o];
                 }
             }