+++ public/vocables/modules/tools/mappingModeler/mappingModelerRelations.js
@@ -3,6 +3,7 @@ import DataSourcesManager from "./dataSourcesManager.js";
 import common from "../../shared/common.js";
 import JstreeWidget from "../../uiWidgets/jstreeWidget.js";
 import MappingModeler from "./mappingModeler.js";
+import UI from "../../shared/UI.js";
 var MappingModelerRelations = (function () {
     self.listPossibleRelations = function () {
         var nodes = MappingColumnsGraph.visjsGraph.data.nodes.get();
@@ -24,7 +25,7 @@ var MappingModelerRelations = (function () {
                 classesMap[nodesMap[edge.to].id] = nodesMap[edge.from].id;
             }
         });
-        var classes =null// Object.keys(classesMap);
+        var classes = null; // Object.keys(classesMap);
         var relations = [];
         Sparql_OWL.getObjectRestrictions(DataSourcesManager.currentSlsvSource, classes, null, function (err, result) {
             result.forEach(function (item) {
@@ -87,6 +88,8 @@ var MappingModelerRelations = (function () {
 
             MappingColumnsGraph.addEdge([edge]);
         });
+        $("#MappingModeler_leftTabs").tabs("option", "active", 1);
+        UIcontroller.onActivateLeftPanelTab("MappingModeler_columnsTab");
     };
 
     return self;